+++ Sources/Extensions/Foundation/NSAttributedStringExtensions.swift
@@ -46,6 +46,7 @@ public extension NSAttributedString {
 
     /// SwifterSwift: Dictionary of the attributes applied across the whole string
     public var attributes: [NSAttributedString.Key: Any] {
+        guard self.length > 0 else { return [:] }
         return attributes(at: 0, effectiveRange: nil)
     }